+++ All.ui/miniprogram/components/colorPicker/colorPicker.js
@@ -45,11 +45,6 @@ Component({
 
   },
 
-  lifetimes: {
-    attached: function() {
-      console.log('美哟？')
-    }
-  },
 
   /**
    * 组件的初始数据