+++ All.ui/miniprogram/pages/index/index.js
@@ -2,6 +2,8 @@
 // 获取应用实例
 const app = getApp()
 import { list } from './config'
+import { changeColor } from '../../utils/util'
+
 Page({
   data: {
     list: list,
@@ -11,6 +13,9 @@ Page({
     wx.getSystemInfo({
       success: (result) => {
         const { statusBarHeight } = result
+
+        // 状态栏高度 + 导航栏高度(44)
+        app.globalData.statusBarHeight = statusBarHeight + 44
         this.setData({
           height: statusBarHeight
         })
@@ -18,10 +23,39 @@ Page({
     })
   },
 
+
   gotoUI: function({ currentTarget }) {
     const { url } = currentTarget.dataset
     wx.navigateTo({
       url,
     })
+  },
+
+  open: function() {
+    this.selectComponent('#sideBox').open()
+  },
+
+
+  onCallbackChange: function(event) {
+    const { detail } = event
+
+    if (detail.color) {
+      const array = this.data.list
+      const list = []
+      for (let i = 0; i < array.length; i++) {
+        let data = {
+          title: array[i].title,
+          list: array[i].list.map(s => {
+            s.svg = changeColor(s.svg, detail.color)
+            return s
+          })
+        }
+        list.push(data)
+      }
+
+      this.setData({
+        list
+      })
+    }
   }
 })